% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotation}
\alias{annotation}
\title{annotation}
\usage{
annotation(
  keys,
  orgDB = "org.Rn.eg.db",
  column = "SYMBOL",
  keytype = "ENSEMBL",
  multi_vals = "first"
)
}
\arguments{
\item{filename}{A character string. The filename.}

\item{ngenes}{An integer. Number of genes to display (default: 1000).}
}
\description{
Helper function to create roc plots
}
