% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{data_input}
\alias{data_input}
\title{data_input}
\usage{
data_input(counttable, metadata, design)
}
\arguments{
\item{counttable}{Filename or path to .csv-file containing raw read counts}

\item{metadata}{Filename or path to .csv-file containing sample descriptions}

\item{design}{Design formula for linear model (as a character string).}
}
\description{
Function for data input and pre-filtering
}
