% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{deg_annotation}
\alias{deg_annotation}
\title{deg_annotation}
\usage{
deg_annotation(keys, orgdb, column, keytype = "ENSEMBL")
}
\arguments{
\item{keys}{existing annotation}

\item{orgdb}{A annotation package}

\item{column}{Data to be retrieved from the database}

\item{keytype}{Type of existing annotation}
}
\description{
Function to add common annotations from
  several common databases
}
