% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the tRomics gui}
\usage{
launch_app(port = 3838, dir = tempdir())
}
\arguments{
\item{port}{The port, tRomics is running on (default: 3838)}

\item{dir}{A character string. The directory, where to create folders to
e.g. store plots etc. (default: tempdir()). CAUTION: relative paths are
not supported.}
}
\value{
tRomics application
}
\description{
Launch the tRomics gui
}
