% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDataimport.R
\name{module_dataimport_ui}
\alias{module_dataimport_ui}
\title{module_dataimport_ui}
\usage{
module_dataimport_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_dataimport_ui
}
