% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleVisualization.R
\name{module_visualization_ui}
\alias{module_visualization_ui}
\title{module_visualization_ui}
\usage{
module_visualization_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_visualization_ui
}
