% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_rocplot}
\usage{
plot_heatmap(rld, filename, ngenes = 1000)
}
\arguments{
\item{filename}{A character string. The filename.}

\item{ngenes}{An integer. Number of genes to display (default: 1000).}
}
\description{
Helper function to create roc plots
}
