% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mds}
\alias{plot_mds}
\title{plot_mds}
\usage{
plot_mds(rld, filename, color_var)
}
\arguments{
\item{filename}{A character string. The filename.}
}
\description{
Helper function to create roc plots
}
