% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pca}
\alias{plot_pca}
\title{plot_pca}
\usage{
plot_pca(rld, filename, color_var)
}
\arguments{
\item{rld}{Object containing log-transformed counts}

\item{filename}{A character string. The filename.}

\item{color_var}{Variable deciding group coloring}
}
\description{
Function for principle component analysis and plotting
}
