% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot_volcano}
\usage{
plot_volcano(results_object, filename, title)
}
\arguments{
\item{results_object}{The resultsfile generated with deg_analysis function}

\item{filename}{A character string. The filename.}

\item{title}{A characte string. The plot title}
}
\description{
Function for plotting the results of differential
  expression analysis as volcano plot
}
