% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{plotVolcano}
\usage{
plotVolcano(results, title, FDR)
}
\arguments{
\item{results}{The resultsfile.}

\item{title}{A characte string. The plot title}
}
\description{
Function for plotting the results of differential expression analysis as volcano plot
}
