% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotation}
\alias{annotation}
\title{annotation}
\usage{
annotation(
  keys,
  orgdb = "org.Rn.eg.db",
  column = "SYMBOL",
  keytype = "ENSEMBL"
)
}
\arguments{
\item{keys}{existing annotation}

\item{column}{Data to be retrieved from the database}

\item{keytype}{Type of existing annotation}

\item{orgdn}{A annotation package}
}
\description{
Function to add common annotations from several common databases
}
