% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_analysis.R
\name{deg_analysis}
\alias{deg_analysis}
\title{deg_analysis}
\usage{
deg_analysis(group_variable, compare_group, control_group)
}
\arguments{
\item{group_variable}{Category of comparison}

\item{compare_group}{attribute group in comparison}

\item{control}{group in comparison}
}
\description{
Function for calculating DEGs and export of results
}
