% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(data, filename, ngenes = 1000)
}
\arguments{
\item{data}{Object containing log-transformed counts}

\item{filename}{A character string. The filename.}

\item{ngenes}{Number of top variance genes to be included in the heatmap}
}
\description{
Function to create heatmap of top-ranked genes (by variance)
}
