% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mds}
\alias{plot_mds}
\title{plot_mds}
\usage{
plot_mds(rld, filename, color_var)
}
\arguments{
\item{rld}{Object containing log-transformed counts}

\item{filename}{A character string. The filename.}

\item{color_var}{Variable deciding group coloring}
}
\description{
Function to create multi-dimensional scaling plot
}
